requirements.txt
langchain==0.1.20
openai
faiss-cpu
PyPDF2
streamlit
reportlab
python-dateutil
langchain-community
python-dotenv